requirements.txt
uwsgi
psycopg2-binary
Django==2.0.7
elasticsearch-dsl==6.0
django-elasticsearch-dsl==0.5.0
djangorestframework==3.8.2
django-elasticsearch-dsl-drf==0.11
autopep8
pylint
ipython
